refactor: 불필요한 ProxyFactory 삭제
